pyproject.toml
[tool.poetry]
name = "repl_python3_bottutorial"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
discord = "^1.7.3"
discord-components = "^2.1.2"
discord-easy-commands = "^0.1.1"
discord-buttons = "^0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
